Trait aptos_parallel_executor::task::Transaction

pub trait Transaction: Sync + Send + 'static {
    type Key: PartialOrd + Send + Sync + Clone + Hash + Eq + ModulePath;
    type Value: Send + Sync + TransactionWrite;
}

Trait that defines a transaction that could be parallel executed by the scheduler. Each transaction will write to a key value storage as their side effect.

Required Associated Types

type Key: PartialOrd + Send + Sync + Clone + Hash + Eq + ModulePath

type Value: Send + Sync + TransactionWrite