Trait aptos_parallel_executor::task::TransactionOutput

pub trait TransactionOutput: Send + Sync {
    type T: Transaction;

    fn get_writes(
        &self
    ) -> Vec<(<Self::T as Transaction>::Key, <Self::T as Transaction>::Value)>;
    fn get_deltas(&self) -> Vec<(<Self::T as Transaction>::Key, DeltaOp)>;
    fn skip_output() -> Self;
}

Trait for execution result of a transaction.

Required Associated Types

type T: Transaction

Type of transaction and its associated key and value.

Required Methods

fn get_writes(
&self
) -> Vec<(<Self::T as Transaction>::Key, <Self::T as Transaction>::Value)>

Get the writes of a transaction from its output.

fn get_deltas(&self) -> Vec<(<Self::T as Transaction>::Key, DeltaOp)>

Get the deltas of a transaction from its output.

fn skip_output() -> Self

Execution output for transactions that comes after SkipRest signal.